requirements.txt
stravalib
google-cloud-storage
google-cloud-secret-manager
